This PR was reverted due to internal test failures, which should be fixed now (and is being sent as an exported diff to make sure).

Original summary:
One less set of independent implementations to worry about going forward
(e.g., we don't have to vectorize these separately from elementwise_util
and they get all benefits of elementwise_util).

Differential Revision: [D76754824](https://our.internmc.facebook.com/intern/diff/D76754824/)

[ghstack-poisoned]

+++ kernels/portable/cpu/util/vectorized_math.h
@@ -151,8 +151,8 @@ inline double reciprocal(double x) {
 template <
     typename Integer,
     std::enable_if_t<std::is_integral_v<Integer>, bool> = true>
-double reciprocal(Integer x) {
-  return reciprocal((double)x);
+float reciprocal(Integer x) {
+  return reciprocal((float)x);
 }
 
 template <
